requirements.txt
matplotlib==3.5.1
numpy==1.22.3
PyYAML==6.0.3
tqdm==4.67.3
